requirements.txt
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
streamlit
pypdf
python-dotenv